requirements.txt



gradio
requests
wikipedia
transformers
torch
pydub
SpeechRecognition
scipy
numpy
soundfile
torchvision
timm
Pillow
huggingface_hub
facenet-pytorch